+++ tracer/src/Datadog.Trace/AppSec/Asm.cs
@@ -1,12 +1,13 @@
-// <copyright file="FeaturesProduct.cs" company="Datadog">
+// <copyright file="Asm.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
-using Datadog.Trace.RemoteConfigurationManagement;
+#nullable enable
 
-namespace Datadog.Trace.Configuration;
-internal class FeaturesProduct : Product
+namespace Datadog.Trace.AppSec;
+
+internal class Asm
 {
-    public override string Name => "FEATURES";
+    public bool Enabled { get; set; }
 }
