+++ tracer/src/Datadog.Trace/AppSec/AsmFeatures.cs
@@ -1,11 +1,15 @@
-// <copyright file="Features.cs" company="Datadog">
+// <copyright file="AsmFeatures.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
-namespace Datadog.Trace.Configuration;
+#nullable enable
 
-internal class Features
+using Datadog.Trace.AppSec;
+
+namespace Datadog.Trace.AppSec;
+
+internal class AsmFeatures
 {
     public Asm Asm { get; set; } = new();
 }
