+++ tracer/src/Datadog.Trace/AppSec/AsmFeaturesProduct.cs
@@ -1,0 +1,14 @@
+// <copyright file="AsmFeaturesProduct.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+#nullable enable
+
+using Datadog.Trace.RemoteConfigurationManagement;
+
+namespace Datadog.Trace.AppSec;
+internal class AsmFeaturesProduct : Product
+{
+    public override string Name => "ASM_FEATURES";
+}
