+++ tracer/src/Datadog.Trace/AppSec/AsmRemoteConfigurationProducts.cs
@@ -1,11 +1,13 @@
-// <copyright file="Asm.cs" company="Datadog">
+// <copyright file="AsmRemoteConfigurationProducts.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
-namespace Datadog.Trace.Configuration;
+#nullable enable
 
-internal class Asm
+namespace Datadog.Trace.AppSec;
+
+internal class AsmRemoteConfigurationProducts
 {
-    public bool Enabled { get; set; }
+    public static AsmFeaturesProduct AsmFeaturesProduct { get; } = new();
 }
