--- tracer/src/Datadog.Trace/Configuration/SharedRemoteConfiguration.cs
@@ -1,18 +1,0 @@
-// <copyright file="SharedRemoteConfiguration.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Datadog.Trace.RemoteConfigurationManagement;
-
-namespace Datadog.Trace.Configuration;
-
-internal class SharedRemoteConfiguration
-{
-    public static FeaturesProduct FeaturesProduct { get; } = new();
-}
